Add "required" task option
Adds a setting field for "required" tasks which can be set when
registering tasks. If a task is registered with "required" => true, then
that task's callback must return true in order for a post to be
transitioned into publish or future status.

If a user attempts to publish a post which does not meet all required
tasks' requirements, the post will remain in its previous status and a
message will be displayed in admin notices alerting the user of the
condition(s) which must be met in order to publish the post.

"Required" tasks are also displayed first in the checklist in bold text
for emphasis.