Evaluate required tasks as part of evaluate_checklist()
Rather than calling the function twice, once for all tasks and once for
required tasks only, it makes sense to just add another key to the
response array. Also changes the notice in the publishing actions meta
box, to only show the notice about submitting for review if the post
isn't published.